v1.5
- The dll injection method now uses `LoadLibraryW` instead of `LoadLibraryA` for better unicode support.
- Added better safety handling and error checks in case something goes wrong.
- Upgraded the project from `.NET 9.0` to `.NET 1.0`.

+++ Forms/MainFrm.cs
@@ -76,7 +76,7 @@ namespace LibraryLoader
         {
             if (e.KeyCode == Keys.Enter)
             {
-                Process? foundProcess = FProcess.GetProcess((Int32)PIDBx.Value);
+                Process? foundProcess = FProcess.GetProcess(PIDBx.Value);
 
                 if (foundProcess != null)
                 {
